refactor: param으로 post할 데이터를 조합하는 ParamManager 생성
